Adjust data structs

+++ source/Attenuator/Header.h
@@ -249,16 +249,16 @@ bool b_a_sync_start = false; // Denotes pack communications have begun.
 struct __attribute__((packed)) STRUCT {
   uint16_t s;
   uint16_t i;
-  uint16_t d1; // Data 1
-  uint16_t d2; // Data 2
+  uint64_t d1; // Data 1
+  uint64_t d2; // Data 2
   uint16_t e;
 } comStruct;
 
 struct __attribute__((packed)) STRUCTSEND {
   uint16_t s;
   uint16_t i;
-  uint16_t d1; // Data 1
-  uint16_t d2; // Data 2
+  uint64_t d1; // Data 1
+  uint64_t d2; // Data 2
   uint16_t e;
 } sendStruct;
