+++ source/ProtonPack/Header.h
@@ -405,8 +405,8 @@ struct __attribute__((packed)) STRUCTSEND {
 struct __attribute__((packed)) STRUCTDATAR {
   uint16_t s;
   uint16_t i;
-  uint16_t d1; // Data 1
-  uint16_t d2; // Data 2
+  uint64_t d1; // Data 1
+  uint64_t d2; // Data 2
   uint16_t e;
 } dataStructR;
 
@@ -414,8 +414,8 @@ struct __attribute__((packed)) STRUCTDATAR {
 struct __attribute__((packed)) STRUCTDATA {
   uint16_t s;
   uint16_t i;
-  uint16_t d1; // Data 1
-  uint16_t d2; // Data 2
+  uint64_t d1; // Data 1
+  uint64_t d2; // Data 2
   uint16_t e;
 } dataStruct;
